# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.BusinessType:
    properties:
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  entity.Category:
    properties:
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
      sub_categories:
        items:
          $ref: '#/definitions/entity.SubCategory'
        type: array
    type: object
...
  entity.Document:
    properties:
      file_type:
        type: string
      filename:
        type: string
      id:
        type: integer
      message:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
  entity.Expertise:
    properties:
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  entity.Image:
    properties:
      id:
        type: integer
      image_url:
        type: string
    required:
    - image_url
    type: object
...
  entity.Location:
    properties:
      address:
        type: string
      formatted_address:
        type: string
      icon:
        type: string
      id:
        type: integer
      international_phone_number:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      phone_number:
        type: string
      place_id:
        type: string
    required:
    - formatted_address
    - lat
    - lng
    type: object
  entity.Occupation:
    properties:
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  entity.Portfolio:
    properties:
      id:
        type: integer
      image_url:
        type: string
      service_id:
        type: integer
      title:
        type: string
    type: object
...
  entity.Rate:
    properties:
      criteria:
        type: string
      height:
        type: integer
      id:
        type: integer
      price:
        type: number
      sub_category_slug:
        type: string
      total_followers:
        type: integer
    type: object
...
  entity.Service:
    properties:
      backgrounds:
        items:
          $ref: '#/definitions/entity.Image'
        type: array
      category:
        $ref: '#/definitions/entity.Category'
        type: object
      category_id:
        type: integer
      cpv:
        type: number
      description:
        type: string
      id:
        type: integer
      language:
        type: string
      portfolios:
        items:
          $ref: '#/definitions/entity.Portfolio'
        type: array
      price:
        type: number
      status:
        type: string
      sub_category:
        $ref: '#/definitions/entity.SubCategory'
        type: object
      sub_category_id:
        type: integer
      title:
        type: string
      topics:
        items:
          $ref: '#/definitions/entity.Expertise'
        type: array
      user:
        $ref: '#/definitions/entity.User'
        type: object
      user_id:
        type: integer
    required:
    - description
    - language
    - title
    type: object
...
  entity.SubCategory:
    properties:
      category_id:
        type: integer
      is_searchable:
        type: boolean
      is_usable:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
  entity.Talent:
    properties:
      background:
        $ref: '#/definitions/entity.Image'
        type: object
      background_image_id:
        type: integer
      birth_date:
        type: string
      business_type_id:
        type: integer
      description:
        type: string
      engagement:
        type: number
      expertises:
        items:
          $ref: '#/definitions/entity.Expertise'
        type: array
      facebook_followers_count:
        type: integer
      facebook_link:
        type: string
      gender:
        type: string
      height:
        type: integer
      id:
        type: integer
      image:
        $ref: '#/definitions/entity.Image'
        type: object
      instagram_followers_count:
        type: integer
      instagram_link:
        type: string
      is_verified:
        type: boolean
      location:
        $ref: '#/definitions/entity.Location'
        type: object
      location_id:
        type: integer
      occupations:
        items:
          $ref: '#/definitions/entity.Occupation'
        type: array
      phone:
        type: string
      twitter_followers_count:
        type: integer
      twitter_link:
        type: string
      youtube_followers_count:
        type: integer
      youtube_link:
        type: string
    required:
    - birth_date
    - gender
    - height
    - instagram_link
    type: object
  entity.TalentResults:
    properties:
      category_name:
        type: string
      image_url:
        type: string
      name:
        type: string
      service_id:
        type: integer
      start_price:
        type: number
      sub_category_name:
        type: string
      talent_id:
        type: integer
    type: object
...
  entity.User:
    properties:
      active:
        type: boolean
      background_image_url:
        type: string
      currency:
        type: string
      email:
        type: string
      image_url:
        type: string
      language:
        type: string
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/entity.Service'
        type: array
      talent:
        $ref: '#/definitions/entity.Talent'
        type: object
      type:
        type: string
    required:
    - currency
    - email
    - language
    - name
    - type
    type: object
...
  translate.ResultErrors:
    properties:
      message:
        example: Dynamic message
        type: string
      status:
        example: Bad Request
        type: string
      status_code:
        example: 400
        type: integer
    type: object
  translate.ResultSuccess:
    properties:
      data:
        type: object
      status:
        example: OK
        type: string
      status_code:
        example: 200
        type: integer
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is event staffing application management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Spgku Application
  version: "1.0"
paths:
...
  /business-type:
    get:
      consumes:
      - application/json
      description: used to find all business types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.BusinessType'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: GetBusinesType used to find all business types
      tags:
      - categories
...
  /campaigns/documents:
    post:
      consumes:
      - multipart/form-data
      description: Upload file
      parameters:
      - description: campaign image
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Document'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/translate.ResultErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: Upload documents
      tags:
      - campaigns
...
  /categories:
    get:
      consumes:
      - application/json
      description: find all category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Category'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: GetCategories used to find all categories
      tags:
      - categories
...
  /expertises:
    get:
      consumes:
      - application/json
      description: used to find all expertises
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Expertise'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: GetExpertises used to find all expertises
      tags:
      - categories
...
  /occupations:
    get:
      consumes:
      - application/json
      description: used to find all occupations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Occupation'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: GetOccupations used to find occupations
      tags:
      - categories
...
  /salary-rates/{id}:
    get:
      consumes:
      - application/json
      description: used to find all salary rates
      parameters:
      - description: SubCategory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Rate'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: GetSalaryRates used to find salary rates
      tags:
      - categories
  /sub-categories:
    get:
      consumes:
      - application/json
      description: find all sub category
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.SubCategory'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: GetSubCategories used to find all sub category
      tags:
      - categories
  /sub-categories/{id}:
    get:
      consumes:
      - application/json
      description: used to find all sub category by category id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.SubCategory'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      summary: GetSubCategoriesByCategoryID used to find all sub category by category id
      tags:
      - categories
  /talent:
    get:
      consumes:
      - application/json
      description: find talent login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Talent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: FindTalentByLogin used to find talent login
      tags:
      - talents
    post:
      description: Create a new talent
      parameters:
      - description: New Talent
        in: body
        name: talent
        required: true
        schema:
          $ref: '#/definitions/entity.Talent'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Talent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: AddTalent api used to create new talent
      tags:
      - talents
  /talent/{id}:
    get:
      consumes:
      - application/json
      description: find talent by id
      parameters:
      - description: Talent ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Talent'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: FindtalentById used to find talent by it's primary key
      tags:
      - talents
  /talent/service:
    post:
      description: Create a new service
      parameters:
      - description: New Service for talent
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/entity.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Service'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: AddService api used to create new service for talent
      tags:
      - talents
  /talent/service/{id}:
    get:
      consumes:
      - application/json
      description: find service by id
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  $ref: '#/definitions/entity.Service'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: FindServiceByID used to find service by primary key
      tags:
      - talents
  /talents:
    get:
      consumes:
      - application/json
      description: find talents
      parameters:
      - in: query
        name: campaign_id
        type: integer
      - in: query
        name: category_id
        type: integer
      - in: query
        name: expertise_name
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: q
        type: string
      - in: query
        name: sub_category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/translate.ResultSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.TalentResults'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/translate.ResultErrors'
      security:
      - ApiKeyAuth: []
      summary: GetTalents is api to find talents by params
      tags:
      - talents
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"